feat: 재고 목록 조회 초기 설계

+++ cs25server/src/main/kotlin/edu/dongguk/cs25server/domain/ItemHQ.kt
@@ -12,10 +12,10 @@ class ItemHQ(
     private var itemName: String,
 
     @Column(name = "price")
-    private var price: Int,
+    private var price: Long,
 
     @Column(name = "stock")
-    private var stock: Int,
+    private var stock: Long,
 
     @Enumerated(EnumType.STRING)
     private var category: ItemCategory
